web/compfest-pay-v2/src/web/server/requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
djangorestframework==3.12.4
Pillow
psycopg2-binary==2.9.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
